pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioredis>=2.0.1",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "python-jose[cryptography]>=3.4.0",
    "sse-starlette>=2.3.5",
    "tortoise-orm[asyncpg]>=0.25.0",
    "uvicorn>=0.34.2",
]
